TestGAN/requirements.txt
asttokens==2.4.1
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
imageio==2.34.2
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.47
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
setuptools==72.1.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
